pyproject.toml
# Copyright Justin R. Goheen.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[tool.poetry]
name = "ibtrader"
version = "0.0.5"
description = "A Python Framework for Interactive Brokers TWS API"
authors = ["Justin R. Goheen"]
license = "Apache-2.0"
readme = "README.md"
homepage="https://github.com/JustinGoheen/ibtrader"

classifiers=[
    "Environment :: Console",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Software Development :: User Interfaces",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
]

include = [
    "*.cff"
]

exclude = [
    "*.toml",
    "requirements.txt",
    "__pycache__",
    "*.lightningai",
    "*.circleci",
    "*.devcontainer",
    "*.github",
    ".gitpod.yml",
]

[tool.poetry.scripts]
trader = "ibquant.cli.console:main"

[tool.poetry.dependencies]
python = ">3.9, <3.11"
ib-insync = "0.9.70"
alphalens-reloaded = "0.4.2"
empyrical-reloaded = "0.5.8"
pyfolio-reloaded = "0.9.4"
zipline-reloaded = "2.2.0"
yfinance = "0.1.72"
quantlib = "1.26"
pyzmq = "23.1.0"
torch = ">=1.12.1"
optuna = ">=2.10.1"
wandb = ">=0.13.1"
torch-tb-profiler = ">=0.4.0"
click = ">=8.1.3"
hydra-core = ">=1.2.0"
textual = "0.1.18"
rich = ">=12.5.1"
rich-cli = ">=1.8.0"
plotly = ">=5.9.0"
matplotlib = ">=3.5.2"
fastapi = "0.79.0"
pyarrow = "8.0.0"
xarray = "^2022.6.0"
pre-commit = "^2.20.0"

[tool.poetry.dev-dependencies]
black = ">=22.1.0"
flake8 = ">=5.0.4"
pytest = ">=7.1.2"
mypy = ">=0.971"
coverage = ">=6.4.3"
pre-commit = ">=2.20.0"
isort = ">=5.10.1"
mkdocs-material = "8.3.9"
mkdocstrings = "0.19.0"


[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[flake8]
max-line-length = 120

[tool.isort]
known_first_party = [
    "ibtrader",
    "ib_insync"
]
profile = "black"
line_length = 120
force_sort_within_sections = "False"
order_by_type = "False"

[tool.mypy]
files = ["lightning_pod"]
install_types = true
non_interactive = true
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
allow_redefinition = true
disable_error_code = "attr-defined"
warn_no_return = "False"

# do not add type hints to lightnig_pod/cli/seed/ or core.module
# because lightning already defines types
[[tool.mypy.overrides]]
module = []
ignore_errors = true

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
norecursedirs = [
    ".git",
    ".github",
    ".circleci",
    "*.egg-info"
]
addopts = "--disable-pytest-warnings"
# filterwarnings = []

[tool.coverage.run]
disable_warnings = ["couldnt-parse"]

[tool.coverage.report]
ignore_errors = true
exclude_lines = ["pragma: no cover"]
